♻️ refactor: baseUrl 의 헤더 함수 수정
