+++ src/constants/globalConstants.ts
@@ -1,1 +1,1 @@
-export const IS_SERVER = typeof window !== 'undefined';
+export const IS_SERVER = typeof window === 'undefined';
