+++ src/utils/cookies.ts
@@ -15,12 +15,12 @@ export function getCookie(
     `(?:^|; )${name.replace(/([.$*|{}()[\]\\/+^])/g, '\\$1')}=([^;]*)`,
   );
 
-  if (cookieString) {
+  if (IS_SERVER && cookieString) {
     const matches = cookieString.match(cookiePattern);
     return matches ? decodeURIComponent(matches[1]) : undefined;
   }
 
-  if (IS_SERVER) {
+  if (cookieString) {
     const matches = document.cookie.match(cookiePattern);
     return matches ? decodeURIComponent(matches[1]) : undefined;
   }
